+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * expose `box` and `unbox` functions
 * add line number to errors [#416](https://github.com/jcubic/lips/issues/416)
 * allow using `define-macro` with `lambda` expression
+* add uninterned symbols SRFI-258
 ### Bugfix
 * fix exception when handling parse error for lonely `)` [#417](https://github.com/jcubic/lips/issues/417)
 * fix unboxing arguments of LIPS created classes [#411](https://github.com/jcubic/lips/issues/411)